refactor(meter-values): unidirectional DAG for coherent split (issue #1936 i)
Address the design findings raised by post-split review by relocating
runtime state and PRNG plumbing to their responsibility owners:

- Move SessionRuntime, sessionRuntimes, getSessionRuntime to
  CoherentSampleComputer.ts as module-private (they exist only to cache
  the voltage-noise PRNG closure across samples, which is a physics
  concern, not a session-lifecycle concern).
- Move disposeCoherentSessionRuntime to CoherentSampleComputer.ts
  (exported for CoherentMeterValuesManager). Update the manager's
  direct import path accordingly.
- Move createStreamPrng to Prng.ts alongside deriveSeed / mulberry32 /
  hashLabel. It composes those three helpers and is a PRNG primitive,
  not a session helper.
- Rename CoherentMeterValuesGenerator.ts to CoherentSession.ts. After
  the moves the file owns only session identity (createCoherentSession,
  CreateSessionOptions), the strategy-gate type guard
  (isCoherentModeActive), and the root-seed resolver (resolveRootSeed).
  The 'Generator' name no longer describes the file — the entry point
  buildCoherentMeterValue lives in the builder.
- Align banner on the renamed file to same-folder convention (Copyright
  ... 2021-2025, matching EvProfiles / Prng / types).
- Update barrel and test imports.

Result: strictly unidirectional import DAG within meter-values/:

  types  ← { CoherentSession, CoherentSampleComputer,
             CoherentMeterValueBuilder, EvProfiles, Prng }
  Prng   ← { CoherentSession, CoherentSampleComputer }
  EvProfiles ← { CoherentSession, CoherentSampleComputer }
  CoherentSampleComputer ← CoherentMeterValueBuilder

CoherentSession no longer depends on any coherent sibling; Computer no
longer depends on Session (the pre-refactor Computer → Session inversion
is eliminated). Three previously-widened intra-package exports revert to
module-private (SessionRuntime, sessionRuntimes, getSessionRuntime);
ROUNDING_SCALE stays exported (Computer → Builder is the correct
direction). Barrel public surface unchanged.

Wire behavior and physics chain byte-identical — the move preserves
every function body verbatim, only relocating them. Test invariant
fail:0, skipped:6 preserved (2925 pass / 2931 total).

The four open PRs touching CoherentMeterValuesGenerator.ts (#1940 c,
#1941 h, #1942 k, #1944 g) will need manual rebase resolution on top
of this refactor — accepted trade-off per the maintainer's directive
to do the design work now rather than defer.
