+++ src/charging-station/meter-values/Prng.ts
@@ -8,8 +8,8 @@
  *
  * Stream splitting: `deriveSeed(rootSeed, label)` XORs a stable FNV-1a
  * 32-bit hash of the label into the root seed so adding one consumer
- * (e.g. a new `POWER_NOISE` stream) does not shift any other stream's
- * sequence.
+ * (e.g. a new `VOLTAGE_NOISE`, `PROFILE_PICK`, or `INITIAL_SOC` stream)
+ * does not shift any other stream's sequence.
  */
 
 /**
@@ -54,13 +54,35 @@ export const hashLabel = (label: string): number => {
  * so two chains collide when `H(x1) ^ H(y1) === H(x2) ^ H(y2)`. Birthday
  * bound on the 32-bit hash space is negligible at simulator scale
  * (expected collisions ≈ N²/2^33; ≈ 0.3 at N = 5×10⁴). The deterministic
- * self-inverse `H(x) ^ H(x) === 0` is neutralized by
- * {@link ./CoherentMeterValuesGenerator.createStreamPrng} namespacing the
- * transactionId leg with a `tx:` prefix labels never carry.
+ * self-inverse `H(x) ^ H(x) === 0` is neutralized by {@link createStreamPrng}
+ * namespacing the transactionId leg with a `tx:` prefix labels never carry.
  * @param rootSeed - Root 32-bit seed.
  * @param label - Stable stream label.
  * @returns Derived 32-bit unsigned seed.
  */
 export const deriveSeed = (rootSeed: number, label: string): number => {
   return ((rootSeed >>> 0) ^ hashLabel(label)) >>> 0
+}
+
+/**
+ * Deterministic per-transaction stream splitter. Combines the station
+ * `randomSeed` (or a stable fallback), the transactionId, and a label so
+ * that adding a new consumer never shifts an existing stream's sequence.
+ * @param rootSeed - Root 32-bit seed for the station.
+ * @param transactionId - Transaction identifier.
+ * @param label - Stream label (`'VOLTAGE_NOISE'`, `'PROFILE_PICK'`,
+ *   `'INITIAL_SOC'`, ...).
+ * @returns PRNG function producing [0, 1) floats.
+ */
+export const createStreamPrng = (
+  rootSeed: number,
+  transactionId: number | string,
+  label: string
+): (() => number) => {
+  // Namespace the transactionId leg with a `tx:` prefix so
+  // `String(transactionId) === label` cannot trigger the XOR self-inverse
+  // `deriveSeed(deriveSeed(r, X), X) === r`. Labels never start with `tx:`
+  // by construction (`VOLTAGE_NOISE`, `PROFILE_PICK`, `INITIAL_SOC`, ...).
+  const txSeed = deriveSeed(rootSeed, `tx:${String(transactionId)}`)
+  return mulberry32(deriveSeed(txSeed, label))
 }
